Update ros_lib generation to include mower_msgs
Need headers for OpenMower msgs to publish on the OpenMower topics

+++ stm32/ros_usbnode/src/ros/generate_ros_lib.sh
@@ -1,8 +1,20 @@
+#! /bin/bash
+set -e
 
-# copy rosserial from our ROS host
-scp -r ubuntu@10.146.111.222:/home/ubuntu/mowgli_ws/libraries/ros_lib ros_lib/
+if ! [ -z "${VIRTUAL_ENV}" ] ; then
+  echo "Do not run from a virtualenv"
+  exit 1
+fi
+
+OPENMOWER_ROS_DIR=${OPENMOWER_ROS_DIR:-~/OpenMower/ROS}
+
+source /opt/ros/noetic/setup.bash
+source ${OPENMOWER_ROS_DIR}/devel/setup.bash
+
+rm -rf ros_lib
+rosrun rosserial_arduino make_libraries.py . mower_msgs
 
 # overwrite arduino code with STM32 code
-cp extra/* ros_lib
+cp extra/* ros_lib/
 rm ros_lib/ArduinoHardware.h
 rm ros_lib/ArduinoTcpHardware.h
